api/inference/requirements.txt
fastapi
uvicorn
torch
numpy
google-cloud-storage
sqlalchemy
psycopg2-binary
